python_service/pyproject.toml
[tool.poetry]
name = "python-service"
version = "0.1.0"
description = "This is a Microservice that allows this project"
authors = ["Caleb Peffer <44934913+calebpeffer@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "python_service"}]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^3.0.2"
llama-index = "^0.10.18"
openai = "^1.13.3"
supabase = "^2.4.0"
llama-index-retrievers-bm25 = "^0.1.3"
flask-cors = "^4.0.0"
gunicorn = "^21.2.0"
llama-index-graph-stores-neo4j = "^0.1.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

